Refactor api.py: extract Calendar routes (CalDAV + Google OAuth)
Continues the api.py split. Stacks on top of #365.

- CalDAV config + status (4 endpoints) → routes/calendar.py
- Google OAuth (status, credentials, auth-url, fetch-token, callback,
  disconnect, direct-auth-url) (7 endpoints) → routes/calendar.py
- Per-agent calendar enable/disable/status (3 endpoints) → routes/calendar.py

Total: 14 endpoints moved.

The OAuth state-nonce CSRF defense logic (`_issue_google_oauth_state` /
`_consume_google_oauth_state`) and the proxied + direct OAuth flows are
self-contained — only `agents` (AgentRegistry) is needed as a dependency,
and the file owns its own `_google_token_store()` constructor.

api.py: 8,816 → 8,389 lines (-427, -4.8%)
Cumulative since main: 11,138 → 8,389 (-2,749 lines, -24.7%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
